backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1
werkzeug==3.1.3
python-dotenv==1.1.0
Pillow==11.1.0
gunicorn==23.0.0
PyJWT==2.10.1
